# ARROW-15320: [Go] Implement memset_neon with Arm64 GoLang Assembly