… go/arrow/memory/Makefile
… go/arrow/memory/memory_neon_arm64.s