@guyuqi 2022-01-14
> Thanks much! @guyuqi this looks great. I don't have easy access to a NEON ARM64 architecture machine, do you think you could the same for the assembly files in https://github.com/apache/arrow/tree/master/go/parquet/internal/utils for the Parquet encoding / bit packing / etc. ?

Absolutely yes. :-)
I'd like to add the simlilar Arm64 Neon implementation in `go/parquet/internal/utils` for Parquet encoding / bit packing.